+++ src/index.ts
@@ -8,7 +8,7 @@ export { useLoading } from './api/useLoading';
 export { getLoading } from './api/getLoading';
 export { connect } from './redux/connect';
 export { useComputed } from './reactive/useComputed';
-export { useDefinedModel } from './model/useDefinedModel';
+export { useDefined, useDefinedModel } from './model/useDefined';
 
 // 入口使用
 export { compose } from 'redux';